SecureCrypt/backend/requirements.txt
Flask
cryptography
pycryptodome
flask-cors
